+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to Virtual Media Folders will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.7] - 2025-12-04
+
+### Added
+- Contextual help tab "Virtual Folders" on Media Library page with usage instructions
+- GitHub repository link in the contextual help sidebar under "For more information"
+
 ## [1.0.6] - 2025-12-04
 
 ### Added